Bust became a command rather than a workflow action
I added the capability to to provide a base path to the file functions,
so from now the bust operation performs in the compiled directory taken
from mimosa config ```watch``` section, therefore it is not needed to
provide the destination directory in the files patterns specified in the
module config ```assetsCacheBust.files```.

Due how I thought, in the fist instance, how to perform the cache bust of
the assets file and after I realised about some internal aspects about
mimosa workflow statges and the options received, the action used in the
'postBuild --> beforePackage' make more sense execute as an independent
command to avoid the problem of bust assets which are been busted in a
previous build and then to design another action to perform it in some
step in the ```build workflow``` in the ```buildFile``` or
```buildExtension``` stage, which in that stages are called for each
file and then we can match the files from this module config file
patterns and change the output file name.

Probably we will have to add a cleaning operation as well.