Added the module to the build/watch worlflow stages
I create the function needed to bust the files on the fly, used by
mimosa ```watch``` and ```build``` (```buildFile``` and
```buildExtension```) workflow stages and steps.

Although it works, I need to add a function to use in ```clean``` workflow
stage and other functions to use in ```update``` and ```remove``` steps
from ```watch``` workflow stage to delete the old file.

I also need to add an option in the module configuration to disable the
module in watch and build, just in case that somebody don't want to use
it