Modules works per file rather than in bulk
The modules performs the operations (rename files and delete and rename
and both operations when the matched assets files are update) per file.

It's needed to fix an issue in clean files with files whose name
contains the splitter character.

I also need to add an option to disable the module