Came back to perform the operations in ```postBuild```
I restored the previous strategy to bust the assets. Again, the module
bust the assets in ```postBuild``` stage after realised that renaming
files before ```prePacakage``` may break some other modules
functionalities, e.g. optimisation modules because they look for the
original file name in the compiled directory.

After I took a look to the other modules, I've realised how to avoid to
launch the module functionality in all the builds, so in the build
launched by ```watch``` command, using the command flags (in our case
```--package``` which is ```isPackage```, in the configuration object
received in ```registration``` mimosa function.

The module, right now, bust the assets if the ```--package``` is passed
to ```watch``` or ```build``` command, but ```clean``` functionality
have to be implemented (working in it) to work in ```postClean`` stage